% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{jl_tonal_dissonance}
\alias{jl_tonal_dissonance}
\title{Tonal dissonance}
\usage{
jl_tonal_dissonance(pc_set)
}
\arguments{
\item{pc_set}{(Numeric vector) Pitch-class set to analyse.}
}
\value{
Integer scalar identifying the chord's consonance rank,
with higher values corresponding to increasing degrees of dissonance.
}
\description{
Computes tonal dissonance using the algorithm
of \insertCite{Johnson-Laird2012;textual}{jl12}.
}
\references{
\insertAllCited{}
}
